Grpc auth

+++ UserServiceGrpc/Repository/SecurityPermissionRepository.cs
@@ -7,7 +7,7 @@ namespace UserServiceGrpc.Repository
     public interface ISecurityPermissionRepository
     {
         Task<List<SecurityPermission>> GetAllAsync();
-        Task<SecurityPermission?> GetByIdAsync(int id);
+        Task<SecurityPermission?> GetByIdAsync(long id);
         Task<SecurityPermission?> CreateAsync(SecurityPermission permission);
         Task<SecurityPermission?> UpdateAsync(SecurityPermission permission);
         Task<bool> DeleteAsync(int id, int modifiedBy);
@@ -36,7 +36,7 @@ namespace UserServiceGrpc.Repository
             }
         }
 
-        public async Task<SecurityPermission?> GetByIdAsync(int id)
+        public async Task<SecurityPermission?> GetByIdAsync(long id)
         {
             try
             {
